[None][refactor] Unify sparse attention framework interfaces
Refactor the sparse attention framework to unify the interface from
attention module to backend to AttentionOp:

- Move DSA prediction (Indexer) from MLA into DSATrtllmAttention via
  predict_sparse_indices() hook, eliminating DSA-specific forward paths
  in MLA (forward_impl_with_dsa, forward_context_dsa, forward_generation_dsa)
- Unify MLA.forward_impl() to handle both standard MLA and DSA through
  the backend's prediction interface
- Extract sparse param preparation into _prepare_sparse_params() helper
  in TrtllmAttention.forward()
- Split monolithic sparse/ files into per-algorithm subdirectories
  (rocket/, dsa/, kernels/) with backward-compatible imports
- Fix naming: DSAtrtllmAttentionMetadata -> DSATrtllmAttentionMetadata
- Clean up C++ AttentionOp sparse flags and add documentation
- Remove redundant mUseSparseAttention setting for MLA case
- Add comprehensive docs to SparseAttentionParams struct
