+++ cpp/tensorrt_llm/kernels/sparseAttentionKernels.h
@@ -27,16 +27,50 @@ TRTLLM_NAMESPACE_BEGIN
 namespace kernels
 {
 
+/// Parameters for sparse attention, controlling which KV tokens/pages each head attends to.
+///
+/// There are two independent sparse index paths:
+///   1. Context phase (sparse_kv_indices/offsets): selects which KV tokens to compact into the
+///      KV cache during prefill, reducing memory footprint.
+///   2. Generation phase (sparse_attn_indices/offsets): selects which KV pages/blocks each head
+///      attends to during token generation, reducing compute.
+///
+/// For sparse MLA (DeepSeek-style), sparse_mla_topk and sparse_mla_kv_cache_pool provide
+/// direct access to the KV cache pool for the dedicated sparse MLA kernel.
 struct SparseAttentionParams
 {
-    int32_t* sparse_kv_indices{nullptr};     // [num_kv_heads, num_sparse_kv_indices]
-    int32_t* sparse_attn_indices{nullptr};   // [num_kv_heads, num_sparse_attn_indices]
-    int32_t* sparse_kv_offsets{nullptr};     // [num_contexts + 1]
-    int32_t* sparse_attn_offsets{nullptr};   // [num_generations + 1]
-    int32_t sparse_mla_topk{0};              // for DSA attention
-    void* sparse_mla_kv_cache_pool{nullptr}; // for DSA attention
+    /// Per-head KV token indices for context-phase KV cache compaction.
+    /// Shape: [num_kv_heads, total_sparse_tokens] (ragged, indexed by sparse_kv_offsets).
+    int32_t* sparse_kv_indices{nullptr};
 
+    /// Per-head KV page/block indices for generation-phase sparse computation.
+    /// Shape: [num_kv_heads, total_sparse_blocks] (ragged, indexed by sparse_attn_offsets).
+    /// Fed to trtllm-gen FMHA as per-head sparse page tables.
+    int32_t* sparse_attn_indices{nullptr};
+
+    /// Batch offsets into sparse_kv_indices. Shape: [num_contexts + 1].
+    /// sparse_kv_indices for request i spans [sparse_kv_offsets[i], sparse_kv_offsets[i+1]).
+    int32_t* sparse_kv_offsets{nullptr};
+
+    /// Batch offsets into sparse_attn_indices. Shape: [num_generations + 1].
+    /// sparse_attn_indices for request i spans [sparse_attn_offsets[i], sparse_attn_offsets[i+1]).
+    int32_t* sparse_attn_offsets{nullptr};
+
+    /// Number of KV tokens each query attends to in sparse MLA (DeepSeek-style).
+    /// When > 0, enables the sparse MLA kernel path.
+    int32_t sparse_mla_topk{0};
+
+    /// Raw pointer to the KV cache memory pool for sparse MLA direct access.
+    /// The sparse MLA kernel bypasses the normal paged KV cache indirection and reads
+    /// KV data directly from this pool using sparse_attn_indices.
+    void* sparse_mla_kv_cache_pool{nullptr};
+
+    /// Granularity of sparse_attn_indices entries: 1 = token-level indices,
+    /// N > 1 = block-level indices where each entry covers N tokens.
     int32_t sparse_attn_indices_block_size{1};
+
+    /// Stride (in elements) between consecutive heads in the sparse_attn_indices tensor.
+    /// When 0, heads are stored contiguously (stride = number of indices per head).
     int32_t sparse_attn_indices_stride{0};
 
     std::string toString() const
