+++ tensorrt_llm/_torch/attention_backend/sparse/__init__.py
@@ -1,7 +1,7 @@
-from .utils import (get_flashinfer_sparse_attn_attention_backend,
-                    get_sparse_attn_kv_cache_manager,
-                    get_trtllm_sparse_attn_attention_backend,
-                    get_vanilla_sparse_attn_attention_backend)
+from .registry import (get_flashinfer_sparse_attn_attention_backend,
+                       get_sparse_attn_kv_cache_manager,
+                       get_trtllm_sparse_attn_attention_backend,
+                       get_vanilla_sparse_attn_attention_backend)
 
 __all__ = [
     "get_sparse_attn_kv_cache_manager",
