+++ tensorrt_llm/_torch/attention_backend/sparse/dsa/cache_manager.py
@@ -1,0 +1,144 @@
+import math
+from typing import TYPE_CHECKING, List, Optional, Union
+
+import torch
+
+import tensorrt_llm
+from tensorrt_llm._torch.pyexecutor.resource_manager import KVCacheManager
+from tensorrt_llm._utils import get_size_in_bytes
+from tensorrt_llm.bindings import DataType
+from tensorrt_llm.bindings.executor import KvCacheConfig
+from tensorrt_llm.bindings.internal.batch_manager import \
+    CacheType as CacheTypeCpp
+from tensorrt_llm.mapping import Mapping
+
+ModelConfig = tensorrt_llm.bindings.ModelConfig
+
+if TYPE_CHECKING:
+    from tensorrt_llm.llmapi.llm_args import (DecodingBaseConfig,
+                                               SparseAttentionConfig)
+
+class DSACacheManager(KVCacheManager):
+
+    def __init__(
+        self,
+        kv_cache_config: KvCacheConfig,
+        kv_cache_type: CacheTypeCpp,
+        *,
+        num_layers: int,
+        num_kv_heads: Union[int, List[Optional[int]]],
+        head_dim: int,
+        tokens_per_block: int,
+        # Note that max_seq_len is not necessarily equal to kv_cache_config.num_tokens.
+        # It's derived from the model's BuildConfig for consistency with the C++ backend.
+        max_seq_len: int,
+        max_batch_size: int,
+        mapping: Mapping,
+        dtype: DataType = DataType.HALF,
+        spec_config: Optional["DecodingBaseConfig"] = None,
+        layer_mask: Optional[List[bool]] = None,
+        max_num_tokens: int = 8192,
+        model_config: Optional[ModelConfig] = None,
+        max_beam_width: int = 1,
+        sparse_attn_config: "SparseAttentionConfig",
+        **kwargs,
+    ) -> None:
+        self.quant_block_size = 128
+        self.index_head_dim = sparse_attn_config.index_head_dim
+
+        super().__init__(
+            kv_cache_config,
+            kv_cache_type,
+            num_layers=num_layers,
+            num_kv_heads=num_kv_heads,
+            head_dim=head_dim,
+            tokens_per_block=tokens_per_block,
+            max_seq_len=max_seq_len,
+            max_batch_size=max_batch_size,
+            mapping=mapping,
+            dtype=dtype,
+            spec_config=spec_config,
+            layer_mask=layer_mask,
+            max_num_tokens=max_num_tokens,
+            model_config=model_config,
+            max_beam_width=max_beam_width,
+            enable_indexer_k_cache=True,
+            indexer_k_cache_quant_block_size=128,
+            indexer_k_cache_index_head_dim=self.index_head_dim,
+            **kwargs,
+        )
+        self.num_blocks = self.blocks_in_primary_pool
+
+        # Indexer K cache pool for DSA attention
+        # Shape: [num_blocks, self.tokens_per_block * (index_head_dim + scale_size)]
+        # Non-interleaved layout: [fp8_tok0 | fp8_tok1 | ... | scale_tok0 | scale_tok1 | ...]
+        # Store FP8-quantized k values from the indexer
+        self.indexer_k_cache_pool_per_layer = [
+            self.get_indexer_k_cache_pool_data(layer_idx)
+            for layer_idx in range(self.num_local_layers)
+        ]
+
+    def get_indexer_k_cache_buffers(self, layer_idx: int):
+        """Get indexer k cache buffer from a specific layer pool."""
+        block_size = self.tokens_per_block
+        per_token_size = self.index_head_dim + self.index_head_dim // self.quant_block_size * 4
+        layer_offset = self.layer_offsets[layer_idx]
+        return self.indexer_k_cache_pool_per_layer[layer_offset].view(
+            self.num_blocks, block_size, 1, per_token_size)
+
+    def shutdown(self):
+        # Clear Python references BEFORE C++ frees the underlying CUDA buffers
+        self.indexer_k_cache_pool_per_layer = []
+        super().shutdown()
+
+    @staticmethod
+    def get_cache_size_per_token(model_config: ModelConfig,
+                                 mapping: Mapping,
+                                 num_layers: Optional[int] = None,
+                                 **kwargs):
+        config = model_config.pretrained_config
+        sparse_attn_config = model_config.sparse_attention_config
+        index_head_dim = sparse_attn_config.index_head_dim
+        quant_block_size = 128
+
+        # get kv cache dtype bytes
+        mem_per_token = 2
+        quant_config = model_config.quant_config
+        if quant_config is not None and quant_config.quant_mode.has_fp8_kv_cache(
+        ):
+            mem_per_token = 1
+
+        # get head dim
+        head_dim = config.kv_lora_rank + config.qk_rope_head_dim
+
+        num_attention_layers = KVCacheManager._resolve_num_attention_layers(
+            model_config, mapping, num_layers)
+        mem_per_token *= num_attention_layers * head_dim
+
+        # 1 for K, others for indexer K cache
+        head_dim_factor = (index_head_dim +
+                           index_head_dim // quant_block_size * 4) / head_dim
+        kv_factor = 1 + head_dim_factor
+        mem_per_token *= kv_factor
+        return mem_per_token
+
+    def get_cache_bytes_per_token(self):
+        # self.kv_factor for K, others for indexer K cache
+        head_dim_factor = (self.index_head_dim + self.index_head_dim //
+                           self.quant_block_size * 4) / self.head_dim
+        kv_factor = self.kv_factor + head_dim_factor
+        cache_size_per_token = math.ceil(
+            kv_factor * sum(self.num_kv_heads_per_layer) * self.head_dim)
+
+        if self.dtype not in (DataType.FP8, DataType.HALF, DataType.BF16,
+                              DataType.FLOAT, DataType.NVFP4):
+            raise ValueError(f'Cannot support {self.dtype} KV cache.')
+
+        cache_size_bytes_per_token = get_size_in_bytes(cache_size_per_token,
+                                                       self.dtype)
+        if self.dtype == DataType.NVFP4:
+            cache_size_bytes_per_token += self.calculate_scaling_factor_size_bytes(
+                cache_size_per_token,
+                quant_vector_size=16,
+                scaling_factor_dtype=DataType.FP8)
+        return cache_size_bytes_per_token
