+++ tensorrt_llm/_torch/attention_backend/sparse/dsa/indexer.py
@@ -1,16 +1,12 @@
 import math
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, List, Optional, Tuple
 
 import torch
 import torch.nn as nn
 
-import tensorrt_llm
-import tensorrt_llm.bindings
 from tensorrt_llm._torch.attention_backend.interface import (
     MLAParams, PositionalEmbeddingParams)
-from tensorrt_llm._torch.attention_backend.trtllm import (
-    TrtllmAttention, TrtllmAttentionMetadata)
 from tensorrt_llm._torch.cute_dsl_utils import IS_CUTLASS_DSL_AVAILABLE
 from tensorrt_llm._torch.distributed.ops import allgather
 from tensorrt_llm._torch.modules.layer_norm import LayerNorm
@@ -18,24 +14,16 @@ from tensorrt_llm._torch.modules.linear import Linear
 from tensorrt_llm._torch.modules.multi_stream_utils import \
     maybe_execute_in_parallel
 from tensorrt_llm._torch.modules.rotary_embedding import RotaryEmbedding
-from tensorrt_llm._torch.pyexecutor.resource_manager import KVCacheManager
 from tensorrt_llm._torch.utils import maybe_compile
-from tensorrt_llm._utils import get_size_in_bytes, get_sm_version, prefer_pinned
-from tensorrt_llm.bindings import DataType
-from tensorrt_llm.bindings.executor import KvCacheConfig
-from tensorrt_llm.bindings.internal.batch_manager import \
-    CacheType as CacheTypeCpp
+from tensorrt_llm._utils import prefer_pinned
 from tensorrt_llm.deep_gemm import (fp8_mqa_logits, fp8_paged_mqa_logits,
                                     get_paged_mqa_logits_metadata)
 from tensorrt_llm.llmapi.llm_args import SparseAttentionConfig
 from tensorrt_llm.logger import logger
-from tensorrt_llm.mapping import Mapping
 from tensorrt_llm.models.modeling_utils import QuantConfig
 
-ModelConfig = tensorrt_llm.bindings.ModelConfig
-
 if TYPE_CHECKING:
-    from tensorrt_llm.llmapi.llm_args import DecodingBaseConfig
+    pass
 
 # Optional import: fast-hadamard-transform causes CI build issues (requires wheel+torch pre-installed)
 try:
@@ -44,7 +32,6 @@ try:
 except ImportError:
     hadamard_transform = None
     HAS_FAST_HADAMARD = False
-
 
 def _compute_slot_mappings(
     global_positions: torch.Tensor,
@@ -291,685 +278,6 @@ class IndexerPrefillChunkMetadata:
     token_end: int  # Q token end index in batch
     k_token_start: int  # K token start index in batch
     k_token_end: int  # K token end index in batch
-
-
-class DSAtrtllmAttentionMetadata(TrtllmAttentionMetadata):
-    # Store reference to indexer for preparation stage
-    indexer: Optional["Indexer"] = None
-    # Chunked prefill metadata for indexer (prefill-only, no CUDA graph needed)
-    indexer_prefill_chunks: Optional[List[IndexerPrefillChunkMetadata]] = None
-    # Max chunk size for two-level chunking:
-    # 1. Request-level: Pack multiple small requests into one chunk (up to indexer_max_chunk_size)
-    # 2. Intra-request: Split large requests into Q-blocks when seq_len > max_chunk_size
-    indexer_max_chunk_size: int
-    # Topk for sparse MLA
-    sparse_mla_topk: int
-    # max number of draft tokens
-    max_draft_tokens: int = 0
-    # Enable indexer skip for short sequences
-    enable_indexer_skip: bool = False
-    # Whether skip the indexer for context requests
-    skip_indexer_for_ctx_reqs: bool = False
-    # Whether skip the indexer for generation requests
-    skip_indexer_for_gen_reqs: bool = False
-    # Whether to use the expanded buffers for MTP support
-    use_expanded_buffers_for_mtp: bool = False
-
-    def __init__(self, *args, **kwargs):
-        self.num_sms = tensorrt_llm.deep_gemm.get_num_sms()
-        # Cached step-invariant values for transform_local_topk_and_prepare_pool_view.
-        # These are recomputed once per step in _ensure_pool_view_cached() and
-        # reused across all layers to avoid redundant Python/CUDA overhead.
-        # Initialized here as plain instance attributes (not class-level
-        # annotations) to stay invisible to dataclass/torch.compile introspection.
-        self._pool_cache_valid = False
-        self._cached_kv_mgr_id = 0
-        self._cached_pool_view = None
-        self._cached_stride_factor = 0
-        self._cached_tokens_per_block = 0
-        self._cached_block_table_ctx = None
-        self._cached_block_table_gen = None
-        self._cached_req_idx_ctx = None
-        self._cached_req_idx_gen = None
-        super().__init__(*args, **kwargs)
-        if self.sparse_attention_config.indexer_max_chunk_size is not None:
-            self.indexer_max_chunk_size = self.sparse_attention_config.indexer_max_chunk_size
-        else:
-            self.indexer_max_chunk_size = 32768  # Default to 32K tokens for the indexer
-
-    def __post_init__(self):
-        super().__post_init__()
-
-        self.sparse_mla_topk = self.sparse_attention_config.index_topk
-        self.enable_indexer_skip = self.sparse_attention_config.skip_indexer_for_short_seqs
-        capture_graph = self.is_cuda_graph
-
-        self.indexer_k_cache_block_offsets = self.get_empty(
-            self.cuda_graph_buffers,
-            [self.max_num_sequences, self.kv_cache_manager.max_blocks_per_seq],
-            cache_name="indexer_k_cache_block_offsets",
-            dtype=torch.int32,
-            capture_graph=capture_graph,
-        )
-        self.host_indexer_k_cache_block_offsets = torch.zeros_like(
-            self.indexer_k_cache_block_offsets,
-            device='cpu',
-            pin_memory=prefer_pinned(),
-        )
-
-        if not self.enable_context_mla_with_cached_kv:
-            self.ctx_cached_token_indptr = self.get_empty(
-                self.cuda_graph_buffers,
-                (self.max_num_requests + 1, ),
-                cache_name="ctx_cached_token_indptr",
-                dtype=torch.int64,
-                capture_graph=capture_graph,
-            )
-            self.host_ctx_cached_token_indptr = torch.zeros_like(
-                self.ctx_cached_token_indptr,
-                device='cpu',
-                pin_memory=prefer_pinned(),
-            )
-            self.ctx_kv_indptr = self.get_empty(
-                self.cuda_graph_buffers,
-                (self.max_num_requests + 1, ),
-                cache_name="ctx_kv_indptr",
-                dtype=torch.int64,
-                capture_graph=capture_graph,
-            )
-            self.host_ctx_kv_indptr = torch.zeros_like(
-                self.ctx_kv_indptr,
-                device='cpu',
-                pin_memory=prefer_pinned(),
-            )
-
-        # Only when MLA chunked prefill is enabled, we need to gather the full KV for indexer's logit computation.
-        # These buffers will be allocated dynamically in Indexer.prepare() based on actual total_kv_len to save memory.
-        if self.enable_context_mla_with_cached_kv:
-            self.slot_mapping_fp8_fullkv = None
-            self.slot_mapping_scale_fullkv = None
-
-        # New generation buffers for dsa
-        self.gen_cached_token_indptr = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.max_num_requests + 1, ),
-            cache_name="gen_cached_token_indptr",
-            dtype=torch.int64,
-            capture_graph=capture_graph,
-        )
-        self.host_gen_cached_token_indptr = torch.zeros_like(
-            self.gen_cached_token_indptr,
-            device='cpu',
-            pin_memory=prefer_pinned(),
-        )
-        self.gen_kv_indptr = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.max_num_requests + 1, ),
-            cache_name="gen_kv_indptr",
-            dtype=torch.int64,
-            capture_graph=capture_graph,
-        )
-        self.host_gen_kv_indptr = torch.zeros_like(
-            self.gen_kv_indptr,
-            device='cpu',
-            pin_memory=prefer_pinned(),
-        )
-        # Indexer metadata
-        # Separate slot mappings for non-interleaved layout (flat byte indices)
-        self.slot_mapping_fp8 = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.max_num_tokens, ),
-            cache_name="slot_mapping_fp8",
-            dtype=torch.int64,
-            capture_graph=capture_graph,
-        )
-        self.host_slot_mapping_fp8 = torch.zeros_like(
-            self.slot_mapping_fp8,
-            device='cpu',
-            pin_memory=prefer_pinned(),
-        )
-        self.slot_mapping_scale = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.max_num_tokens, ),
-            cache_name="slot_mapping_scale",
-            dtype=torch.int64,
-            capture_graph=capture_graph,
-        )
-        self.host_slot_mapping_scale = torch.zeros_like(
-            self.slot_mapping_scale,
-            device='cpu',
-            pin_memory=prefer_pinned(),
-        )
-        # Per-token request index buffer for topk_indices conversion
-        self.req_idx_per_token = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.max_num_tokens, ),
-            cache_name="req_idx_per_token",
-            dtype=torch.int32,
-            capture_graph=capture_graph,
-        )
-        # Block table for topk_indices conversion (shared for context and generation)
-        self.block_table = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.max_num_requests, self.kv_cache_manager.max_blocks_per_seq),
-            cache_name="block_table",
-            dtype=torch.int32,
-            capture_graph=capture_graph,
-        )
-        self.scheduler_metadata_buffer = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.num_sms + 1, 2),
-            cache_name="scheduler_metadata_buffer",
-            dtype=torch.int32,
-            capture_graph=capture_graph,
-        )
-        self.cu_seqlen_ks = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.max_num_tokens, ),
-            cache_name="cu_seqlen_ks",
-            dtype=torch.int32,
-            capture_graph=capture_graph,
-        )
-        self.cu_seqlen_ke = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.max_num_tokens, ),
-            cache_name="cu_seqlen_ke",
-            dtype=torch.int32,
-            capture_graph=capture_graph,
-        )
-        # Topk indices buffer to support skip indexer for requests with short sequence lengths
-        if self.enable_indexer_skip:
-            self.topk_indices_buffer = self.get_empty(
-                self.cuda_graph_buffers,
-                (self.max_num_tokens, self.sparse_mla_topk),
-                cache_name="topk_indices_buffer",
-                dtype=torch.int32,
-                capture_graph=capture_graph,
-            )
-            self.host_topk_indices_buffer = torch.zeros_like(
-                self.topk_indices_buffer,
-                device='cpu',
-                pin_memory=prefer_pinned(),
-            )
-        # Create expanded buffers for MTP support
-        self.create_expanded_buffers(capture_graph=capture_graph)
-
-    # TODO: remove these expanded buffers when fp8_paged_mqa_logits supports an arbitrary number of MTP draft tokens.
-    def create_expanded_buffers(self, capture_graph=False):
-        self.kv_lens_expanded_cuda = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.max_num_sequences * (1 + self.max_draft_tokens), ),
-            cache_name="kv_lens_expanded_cuda",
-            dtype=torch.int32,
-            capture_graph=capture_graph,
-        )
-        self.kv_lens_expanded_host = torch.zeros_like(
-            self.kv_lens_expanded_cuda,
-            device='cpu',
-            pin_memory=prefer_pinned(),
-        )
-        self.block_table_expanded = self.get_empty(
-            self.cuda_graph_buffers,
-            [
-                self.max_num_sequences * (1 + self.max_draft_tokens),
-                self.kv_cache_manager.max_blocks_per_seq
-            ],
-            cache_name="block_table_expanded",
-            dtype=torch.int32,
-            capture_graph=capture_graph,
-        )
-        self.host_block_table_expanded = torch.zeros_like(
-            self.block_table_expanded,
-            device='cpu',
-            pin_memory=prefer_pinned(),
-        )
-        self.scheduler_metadata_buffer_expanded = self.get_empty(
-            self.cuda_graph_buffers,
-            (self.num_sms + 1, 2),
-            cache_name="scheduler_metadata_buffer_expanded",
-            dtype=torch.int32,
-            capture_graph=capture_graph,
-        )
-        # The fp8_paged_mqa_logits kernel needs different layout of the metadata buffer for MTP=3.
-        if self.max_draft_tokens == 3:
-            self.scheduler_metadata_buffer_mtp3 = self.get_empty(
-                self.cuda_graph_buffers,
-                (self.num_sms // 2 + 1, 2),
-                cache_name="scheduler_metadata_buffer_mtp3",
-                dtype=torch.int32,
-                capture_graph=capture_graph,
-            )
-
-    # This function is only used to create the expanded buffers when the max_draft_tokens is changed.
-    # TODO: remove this function once fp8_paged_mqa_logits supports an arbitrary number of MTP draft tokens.
-    def update_spec_dec_param(
-        self,
-        batch_size,
-        is_spec_decoding_enabled,
-        is_spec_dec_tree,
-        is_spec_dec_dynamic_tree,
-        max_draft_len,
-        max_total_draft_tokens,
-        model_is_wrapped: bool = False,
-        spec_metadata: Optional['SpecMetadata'] = None,
-        spec_tree_manager: Optional['SpecTreeManager'] = None,
-        spec_decoding_tensor: Optional['SpecDecodingTensor'] = None,
-    ):
-        super().update_spec_dec_param(batch_size, is_spec_decoding_enabled,
-                                      is_spec_dec_tree,
-                                      is_spec_dec_dynamic_tree, max_draft_len,
-                                      max_total_draft_tokens, model_is_wrapped,
-                                      spec_metadata, spec_tree_manager,
-                                      spec_decoding_tensor)
-        self.max_draft_tokens = max_draft_len
-        init_shape = self.kv_lens_expanded_host.shape[0]
-        if self.max_num_sequences * (1 + self.max_draft_tokens) != init_shape:
-            capture_graph = self.is_cuda_graph
-            self.create_expanded_buffers(capture_graph=capture_graph)
-
-    def _invalidate_pool_view_cache(self):
-        """Invalidate the cached pool view and related step-invariant values.
-
-        Must be called at the start of each forward step (in prepare()) so that
-        _ensure_pool_view_cached() recomputes them for the new batch.
-        """
-        self._pool_cache_valid = False
-
-    def _ensure_pool_view_cached(self):
-        """Compute and cache values used by
-        transform_local_topk_and_prepare_pool_view().
-
-        These values (pool view, stride factor, block table slices, request
-        index slices) are constant across all layers sharing the same KV pool
-        and batch dimensions within a forward pass. Caching them avoids
-        redundant Python/CUDA overhead per layer.
-
-        Safety: _invalidate_pool_view_cache() is called unconditionally at the
-        start of every step (prepare() and on_update_kv_lens()), so the boolean
-        flag is always cleared before the first per-layer call within a step.
-        """
-        if self._pool_cache_valid and self._cached_kv_mgr_id == id(
-                self.kv_cache_manager):
-            return
-
-        pool = self.kv_cache_manager.get_unique_primary_pool()
-        kv_cache_manager = self.kv_cache_manager
-        num_blocks, num_layers, _, _ = pool.shape
-        self._cached_tokens_per_block = kv_cache_manager.tokens_per_block
-        head_dim = kv_cache_manager.head_dim
-        self._cached_pool_view = pool.squeeze(2).view(-1, 1, head_dim)
-        self._cached_stride_factor = (num_layers *
-                                      self._cached_tokens_per_block)
-        self._cached_block_table_ctx = self.block_table[:self.num_contexts]
-        self._cached_block_table_gen = self.block_table[self.num_contexts:self.
-                                                        num_seqs]
-        self._cached_req_idx_ctx = self.req_idx_per_token[:self.num_ctx_tokens]
-        self._cached_req_idx_gen = (
-            self.req_idx_per_token[self.num_ctx_tokens:self.num_tokens] -
-            self.num_contexts)
-        self._cached_kv_mgr_id = id(kv_cache_manager)
-        self._pool_cache_valid = True
-
-    @maybe_compile(dynamic=True)
-    def _get_dense_topk_indices(self, seq_lens, kv_lens, num_tokens):
-        device = kv_lens.device
-        past_kv_lens = kv_lens - seq_lens
-        # get position ids
-        seq_ends = torch.cumsum(seq_lens, dim=0)
-        seq_starts = seq_ends - seq_lens
-        per_seq_offsets = past_kv_lens - seq_starts  # Shape: [batch_size]
-        global_indices = torch.arange(num_tokens, device=device)
-        batch_indices = torch.searchsorted(seq_ends,
-                                           global_indices,
-                                           side='right')
-        repeated_offsets = per_seq_offsets[batch_indices]
-        position_ids = global_indices + repeated_offsets
-        # get the dense topk indices with causal mask
-        range_row = torch.arange(self.sparse_mla_topk, device=device)
-        mask = range_row <= position_ids.unsqueeze(1)
-        return torch.where(mask, range_row, -1)
-
-    def prepare_dense_topk_indices(self,
-                                   kv_lens,
-                                   device=False):  # device=False means use CPU
-
-        if self.num_contexts > 0 and self.skip_indexer_for_ctx_reqs:
-            ctx_range = slice(self.num_ctx_tokens)
-            if device:
-                self.topk_indices_buffer[ctx_range, :].copy_(
-                    self._get_dense_topk_indices(
-                        self.seq_lens_cuda[:self.num_contexts],
-                        kv_lens[:self.num_contexts], self.num_ctx_tokens),
-                    non_blocking=True)
-            else:
-                self.host_topk_indices_buffer[
-                    ctx_range, :] = self._get_dense_topk_indices(
-                        self.seq_lens[:self.num_contexts],
-                        kv_lens[:self.num_contexts], self.num_ctx_tokens)
-                self.topk_indices_buffer[ctx_range, :].copy_(
-                    self.host_topk_indices_buffer[ctx_range, :],
-                    non_blocking=True)
-
-        if self.num_generations > 0 and self.skip_indexer_for_gen_reqs:
-            gen_range = slice(self.num_ctx_tokens, self.num_tokens)
-            if device:
-                self.topk_indices_buffer[gen_range, :].copy_(
-                    self._get_dense_topk_indices(
-                        self.seq_lens_cuda[self.num_contexts:self.num_seqs],
-                        kv_lens[self.num_contexts:self.num_seqs],
-                        self.num_tokens - self.num_ctx_tokens),
-                    non_blocking=True)
-            else:
-                self.host_topk_indices_buffer[
-                    gen_range, :] = self._get_dense_topk_indices(
-                        self.seq_lens[self.num_contexts:self.num_seqs],
-                        kv_lens[self.num_contexts:self.num_seqs],
-                        self.num_tokens - self.num_ctx_tokens)
-                self.topk_indices_buffer[gen_range, :].copy_(
-                    self.host_topk_indices_buffer[gen_range, :],
-                    non_blocking=True)
-
-    def _get_pool_block_indices(self) -> torch.Tensor:
-        """Extract memory pool block indices from host_kv_cache_block_offsets.
-
-        The C++ setOffsets() encodes offsets as:
-            encoded = memPoolBlockIndex * numLayers * kvFactor
-        For SELFKONLY (MLA/DSA), kvFactor=1, so:
-            memPoolBlockIndex = encoded // num_local_layers
-
-        Returns a (num_seqs, max_blocks_per_seq) int32 CPU tensor with valid
-        pool indices clamped to [0, blocks_in_primary_pool - 1].
-        """
-        num_local_layers = self.kv_cache_manager.num_local_layers
-        max_pool_idx = self.kv_cache_manager.blocks_in_primary_pool - 1
-        # DSA uses SELFKONLY mode where only key cache is stored (kv_factor=1).
-        # host_kv_cache_block_offsets shape: (num_pools, max_batch*beam, 2, max_blocks_per_seq)
-        # Note: dim=2 is always 2 in the tensor layout (K and V slots), but for
-        # SELFKONLY only the K slot (index 0) contains valid data.
-        assert self.kv_cache_manager.kv_factor == 1, \
-            f"DSA requires SELFKONLY mode (kv_factor=1), got kv_factor={self.kv_cache_manager.kv_factor}"
-        # Pool 0, first num_seqs entries, field 0 (key offsets)
-        encoded = self.kv_cache_manager.host_kv_cache_block_offsets[
-            0, :self.num_seqs, 0, :]
-        pool_indices = encoded // num_local_layers
-        # Clamp for safety: handles garbage padding from torch.empty in uninitialized slots
-        pool_indices = pool_indices.clamp(min=0,
-                                          max=max_pool_idx).to(torch.int32)
-        return pool_indices
-
-    def prepare(self):
-        super().prepare()
-        self._invalidate_pool_view_cache()
-
-        # Get kv lengths
-        assert self.kv_cache_params.use_cache is True, "DSA requires use_cache to be True"
-        cached_token_lens = torch.tensor(
-            self.kv_cache_params.num_cached_tokens_per_seq,
-            dtype=torch.int,
-            device='cpu',
-        )
-        if self.enable_helix:
-            # For Helix CP, inactive ranks only attend to previously cached
-            # tokens (no new token appended), while active ranks add new tokens.
-            # This mirrors the kv_lens logic in TrtllmAttentionMetadata.prepare().
-            active_rank = ~self.helix_is_inactive_rank_cpu[:self.num_seqs]
-            kv_lens = cached_token_lens.clone()
-            kv_lens[active_rank] += self.seq_lens_kv[active_rank]
-        else:
-            kv_lens = cached_token_lens + self.seq_lens_kv
-
-        # Prepare to support skip indexer
-        num_extra_kv_tokens = self.kv_cache_params.num_extra_kv_tokens
-        if self.num_contexts > 0 and self.enable_indexer_skip:
-            # Minus the number of extra KV tokens because when using one-model MTP, the
-            # draft layers needs more KV tokens for the next draft forwards.
-            self.skip_indexer_for_ctx_reqs = kv_lens[:self.num_contexts].max(
-            ).item() <= self.sparse_mla_topk - num_extra_kv_tokens
-        else:
-            self.skip_indexer_for_ctx_reqs = False
-
-        if self.num_generations > 0 and self.enable_indexer_skip:
-            # Minus the number of extra KV tokens because when using one-model MTP, the
-            # draft layers needs more KV tokens for the next draft forwards.
-            self.skip_indexer_for_gen_reqs = kv_lens[
-                self.num_contexts:self.num_seqs].max().item(
-                ) <= self.sparse_mla_topk - num_extra_kv_tokens
-        else:
-            self.skip_indexer_for_gen_reqs = False
-        self.prepare_dense_topk_indices(kv_lens)
-
-        # Build indexer_k_cache_block_offsets using pool block indices derived
-        # from host_kv_cache_block_offsets (populated by super().prepare()).
-        # This correctly resolves block IDs to memory pool indices, which is
-        # required when host cache offload is enabled (block IDs != pool indices
-        # for onboarded secondary blocks).
-        if self.kv_cache_manager is not None:
-            pool_indices = self._get_pool_block_indices()
-            self.host_indexer_k_cache_block_offsets[:self.num_seqs].copy_(
-                pool_indices)
-            self.indexer_k_cache_block_offsets[:self.num_seqs].copy_(
-                self.host_indexer_k_cache_block_offsets[:self.num_seqs],
-                non_blocking=True)
-            # Safety clamp: prevent OOB from CUDA graph padding entries which
-            # may contain stale negative or out-of-range values after block
-            # eviction/onboarding with host cache offload.
-            self.indexer_k_cache_block_offsets.clamp_(min=0)
-
-        # Build req_idx_per_token for topk_indices conversion
-        host_req_idx_per_token = torch.repeat_interleave(torch.arange(
-            self.num_seqs, dtype=torch.int32),
-                                                         self.seq_lens,
-                                                         dim=0)
-        self.req_idx_per_token[:self.num_tokens].copy_(host_req_idx_per_token,
-                                                       non_blocking=True)
-
-        # Build block_table for topk_indices conversion (actual block allocation)
-        if self.kv_cache_manager is not None:
-            tokens_per_block = self.kv_cache_manager.tokens_per_block
-            num_blocks_per_seq = (kv_lens[:self.num_seqs] + tokens_per_block -
-                                  1) // tokens_per_block
-            max_blocks_used = num_blocks_per_seq.max().item(
-            ) if self.num_seqs > 0 else 1
-            # pool_indices already has correct values; set padding to -1
-            host_block_table = pool_indices[:, :max_blocks_used].clone()
-            for i in range(self.num_seqs):
-                if num_blocks_per_seq[i] < max_blocks_used:
-                    host_block_table[i, num_blocks_per_seq[i]:] = -1
-            # Copy to GPU
-            self.block_table[:self.num_seqs, :max_blocks_used].copy_(
-                host_block_table, non_blocking=True)
-
-        # For mla_rope_append_paged_kv_assign_q
-        if self.num_contexts > 0:
-            self.num_ctx_cached_tokens = cached_token_lens[:self.
-                                                           num_contexts].sum(
-                                                           ).item()
-            self.max_ctx_kv_len = kv_lens[:self.num_contexts].max().item()
-            self.max_ctx_seq_len = self.seq_lens[:self.num_contexts].max().item(
-            )
-            # context cached token indptr
-            torch.cumsum(
-                cached_token_lens[:self.num_contexts],
-                dim=0,
-                dtype=torch.int64,
-                out=self.host_ctx_cached_token_indptr[1:self.num_contexts + 1])
-            self.ctx_cached_token_indptr[:self.num_contexts + 1].copy_(
-                self.host_ctx_cached_token_indptr[:self.num_contexts + 1],
-                non_blocking=True)
-            # context kv indptr
-            torch.cumsum(kv_lens[:self.num_contexts],
-                         dim=0,
-                         dtype=torch.int64,
-                         out=self.host_ctx_kv_indptr[1:self.num_contexts + 1])
-            self.ctx_kv_indptr[:self.num_contexts + 1].copy_(
-                self.host_ctx_kv_indptr[:self.num_contexts + 1],
-                non_blocking=True)
-        else:
-            self.num_ctx_cached_tokens = 0
-            self.max_ctx_kv_len = 0
-            self.max_ctx_seq_len = 0
-
-        if self.num_generations > 0:
-            self.max_gen_seq_len = self.seq_lens[self.num_contexts:self.
-                                                 num_seqs].max().item()
-            # generation cached token indptr
-            torch.cumsum(
-                cached_token_lens[self.num_contexts:self.num_seqs],
-                dim=0,
-                dtype=torch.int64,
-                out=self.host_gen_cached_token_indptr[1:self.num_generations +
-                                                      1])
-            self.gen_cached_token_indptr[:self.num_generations + 1].copy_(
-                self.host_gen_cached_token_indptr[:self.num_generations + 1],
-                non_blocking=True)
-            # generation kv indptr
-            torch.cumsum(kv_lens[self.num_contexts:self.num_seqs],
-                         dim=0,
-                         dtype=torch.int64,
-                         out=self.host_gen_kv_indptr[1:self.num_generations +
-                                                     1])
-            self.gen_kv_indptr[:self.num_generations + 1].copy_(
-                self.host_gen_kv_indptr[:self.num_generations + 1],
-                non_blocking=True)
-        else:
-            self.max_gen_seq_len = 0
-
-        # Because the fp8_paged_mqa_logits only supports seq_len == 1/2/4 (i.e., max_draft_tokens == 0/1/3) on sm100, and
-        # seq_len == 1/2 (i.e., max_draft_tokens == 0/1) on sm90, for other cases, we need to flatten the q tensor and
-        # expand the kv_lens and block_table for MTP support.
-        # TODO:
-        # - No distinction between sm90 and sm100 is needed once MTP3 is supported on sm90.
-        # - Remove this once fp8_paged_mqa_logits supports an arbitrary number of MTP draft tokens.
-        self.use_expanded_buffers_for_mtp = (
-            (self.max_draft_tokens > 1 and get_sm_version() == 90)
-            or ((self.max_draft_tokens == 2 or self.max_draft_tokens > 3)
-                and get_sm_version() >= 100))
-        if self.use_expanded_buffers_for_mtp:
-            # Expand kv_lens_cuda (only generation)
-            num_tokens = self.num_generations * (1 + self.max_draft_tokens)
-            gen_kv_lens = kv_lens[self.num_contexts:self.num_seqs]
-            gen_kv_lens_expanded = torch.stack([gen_kv_lens] *
-                                               (1 + self.max_draft_tokens),
-                                               dim=0)
-            gen_kv_lens_expanded = gen_kv_lens_expanded.transpose(
-                0, 1).contiguous().flatten()
-            self.kv_lens_expanded_host[:num_tokens].copy_(gen_kv_lens_expanded)
-            self.kv_lens_expanded_cuda[:num_tokens].copy_(
-                self.kv_lens_expanded_host[:num_tokens], non_blocking=True)
-
-            # Expand indexer_k_cache_block_offsets (only generation)
-            # host_indexer_k_cache_block_offsets already contains correct pool
-            # indices from _get_pool_block_indices() above.
-            if self.kv_cache_manager is not None and self.num_generations > 0:
-                max_len = self.host_indexer_k_cache_block_offsets.shape[1]
-                gen_block_tensor = self.host_indexer_k_cache_block_offsets[
-                    self.num_contexts:self.num_seqs, :max_len]
-                expanded_blocks = gen_block_tensor.repeat_interleave(
-                    1 + self.max_draft_tokens, dim=0)
-                self.host_block_table_expanded[:num_tokens, :max_len].copy_(
-                    expanded_blocks, non_blocking=True)
-                self.block_table_expanded[:num_tokens].copy_(
-                    self.host_block_table_expanded[:num_tokens],
-                    non_blocking=True)
-                self.block_table_expanded.clamp_(min=0)
-
-        # Prepare metadata for indexer
-        Indexer.prepare(metadata=self)
-
-    def on_update_kv_lens(self):
-        # After changing the kv_lens/kv_lens_cuda, we may need to update other metadatas.
-        # Especially for the changes in the _preprocess_inputs() of model_engine.py.
-        #
-        # NOTE:
-        # In overlap scheduler + speculative decoding, kv_lens_cuda can be corrected at runtime
-        # (inside _preprocess_inputs) to account for variable accepted tokens. The indexer
-        # slot_mapping_* buffers also depend on these effective cached lengths. If we do not
-        # refresh slot mappings here, indexer K-cache updates can be written with stale offsets.
-
-        # _preprocess_inputs() also uses this as a general hook to "invalidate per-forward-pass
-        # caches so they are recomputed (and captured) on every _forward_step". Invalidate the
-        # pool_view cache here so it is recomputed on the next
-        # transform_local_topk_and_prepare_pool_view() call.
-        self._invalidate_pool_view_cache()
-
-        if self.kv_cache_manager is not None and self.num_tokens > 0:
-            seq_lens = self.seq_lens_cuda[:self.num_seqs]
-            # Runtime cached lengths after overlap/spec-dec correction.
-            start_positions = self.kv_lens_cuda[:self.num_seqs] - seq_lens
-
-            # Reuse request-per-token mapping prepared in metadata.prepare().
-            # This avoids repeat_interleave in graph-capture mode.
-            req_indices = self.req_idx_per_token[:self.num_tokens].to(
-                dtype=torch.int64)
-            seq_starts = torch.cumsum(
-                seq_lens, dim=0, dtype=torch.int64) - seq_lens.to(torch.int64)
-            token_offsets = torch.arange(
-                self.num_tokens, device=seq_lens.device,
-                dtype=torch.int64) - seq_starts[req_indices]
-
-            global_positions = start_positions[req_indices] + token_offsets
-            fp8_indices, scale_indices = _compute_slot_mappings(
-                global_positions,
-                self.indexer_k_cache_block_offsets,
-                req_indices,
-                self.kv_cache_manager.index_head_dim,
-                self.kv_cache_manager.tokens_per_block,
-                self.kv_cache_manager.quant_block_size,
-            )
-            self.slot_mapping_fp8[:self.num_tokens] = fp8_indices
-            self.slot_mapping_scale[:self.num_tokens] = scale_indices
-
-        if self.num_generations > 0:
-            torch.cumsum(
-                self.kv_lens_cuda[self.num_contexts:self.
-                                  num_seqs],  # num_contexts should be 0
-                dim=0,
-                dtype=torch.int64,
-                out=self.gen_kv_indptr[1:self.num_generations + 1])
-            torch.cumsum(
-                (self.kv_lens_cuda[self.num_contexts:self.num_seqs] -
-                 self.seq_lens_cuda[self.num_contexts:self.num_seqs]),
-                dim=0,
-                dtype=torch.int64,
-                out=self.gen_cached_token_indptr[1:self.num_generations + 1])
-            scheduler_metadata_buffer = get_paged_mqa_logits_metadata(
-                self.kv_lens_cuda[self.num_contexts:self.num_seqs],
-                self.kv_cache_manager.tokens_per_block, self.num_sms)
-            self.scheduler_metadata_buffer.copy_(scheduler_metadata_buffer,
-                                                 non_blocking=True)
-            if self.use_expanded_buffers_for_mtp:
-                num_draft_tokens = 1 + self.max_draft_tokens
-                num_tokens = self.num_generations * num_draft_tokens
-                gen_kv_lens = self.kv_lens_cuda[self.num_contexts:self.num_seqs]
-                kv_lens_expanded = torch.stack([gen_kv_lens] * num_draft_tokens,
-                                               dim=0)
-                self.kv_lens_expanded_cuda[:num_tokens] = \
-                    kv_lens_expanded.transpose(0, 1).contiguous().flatten()
-                # Expand schedule metadata buffer (only generation)
-                kv_lens_expanded = self.kv_lens_expanded_cuda[:num_tokens]
-                scheduler_metadata_buffer_expanded = get_paged_mqa_logits_metadata(
-                    kv_lens_expanded, self.kv_cache_manager.tokens_per_block,
-                    self.num_sms)
-                self.scheduler_metadata_buffer_expanded.copy_(
-                    scheduler_metadata_buffer_expanded, non_blocking=True)
-            elif self.max_draft_tokens == 3:
-                scheduler_metadata_buffer_mtp3 = get_paged_mqa_logits_metadata(
-                    self.kv_lens_cuda[self.num_contexts:self.num_seqs],
-                    self.kv_cache_manager.tokens_per_block, self.num_sms // 2)
-                self.scheduler_metadata_buffer_mtp3.copy_(
-                    scheduler_metadata_buffer_mtp3, non_blocking=True)
-        self.prepare_dense_topk_indices(self.kv_lens_cuda, device=True)
-
-    def update_for_spec_dec(self):
-        super().update_for_spec_dec()
-        # host
-        self.max_ctx_kv_len = 0
-        self.num_ctx_cached_tokens = 0
-        self.max_gen_seq_len = 1
-
-        # device
-        self.on_update_kv_lens()
-
 
 @maybe_compile(dynamic=True)
 def _scale(weights: torch.Tensor, q_scale: torch.Tensor,
@@ -1756,256 +1064,3 @@ class Indexer(nn.Module):
         return self.sparse_attn_indexer(metadata, hidden_states, q_fp8, k_fp8,
                                         k_scale, weights)
 
-
-class DSATrtllmAttention(TrtllmAttention):
-    Metadata = DSAtrtllmAttentionMetadata
-
-    def __init__(
-            self,
-            layer_idx: int,
-            num_heads: int,
-            head_dim: int,
-            num_kv_heads: Optional[int] = None,
-            quant_config: Optional[QuantConfig] = None,
-            q_scaling: Optional[float] = None,
-            pos_embd_params: Optional[PositionalEmbeddingParams] = None,
-            mla_params: Optional[MLAParams] = None,
-            skip_create_weights_in_init: bool = False,
-            attention_chunk_size: Optional[int] = None,
-            sparse_attention_config: Optional["SparseAttentionConfig"] = None,
-            dtype: Optional[torch.dtype] = None,
-            aux_stream: Optional[torch.cuda.Stream] = None,
-            **kwargs):
-        if sparse_attention_config is None:
-            raise ValueError(
-                "sparse_attention_config is required for DSATrtllmAttention and cannot be None"
-            )
-        TrtllmAttention.__init__(
-            self,
-            layer_idx,
-            num_heads,
-            head_dim,
-            sparse_attention_config=sparse_attention_config,
-            num_kv_heads=num_kv_heads,
-            quant_config=quant_config,
-            q_scaling=q_scaling,
-            pos_embd_params=pos_embd_params,
-            mla_params=mla_params,
-            skip_create_weights_in_init=skip_create_weights_in_init,
-            attention_chunk_size=attention_chunk_size,
-            **kwargs)
-
-        self.indexer = Indexer(quant_config, pos_embd_params, mla_params,
-                               skip_create_weights_in_init,
-                               sparse_attention_config, dtype, layer_idx,
-                               aux_stream)
-
-    def sparse_attn_predict(
-        self,
-        q: torch.Tensor,
-        k: Optional[torch.Tensor],
-        metadata: DSAtrtllmAttentionMetadata,
-        hidden_states: Optional[torch.Tensor] = None,
-        qr: Optional[torch.Tensor] = None,
-        position_ids: Optional[torch.Tensor] = None,
-        topk_indices: Optional[torch.Tensor] = None,
-        is_generation: bool = True,
-        **kwargs,
-    ) -> Tuple[Optional[torch.Tensor], Optional[torch.Tensor]]:
-        # Transform the local topk indices to global topk indices in paged kv cache
-        topk_indices_global, _ = transform_local_topk_and_prepare_pool_view(
-            topk_indices, metadata, self.get_local_layer_idx(metadata),
-            is_generation)
-
-        # TODO: Use sparse_attn_indexer to predict the indices for DSA attention
-        # return self.indexer(q, k, metadata, hidden_states, qr, position_ids)
-        return topk_indices_global, None
-
-    def sparse_kv_predict(
-        self,
-        q: torch.Tensor,
-        k: Optional[torch.Tensor],
-        metadata: DSAtrtllmAttentionMetadata,
-        hidden_states: Optional[torch.Tensor] = None,
-        qr: Optional[torch.Tensor] = None,
-        **kwargs,
-    ) -> Tuple[Optional[torch.Tensor], Optional[torch.Tensor]]:
-        return None, None
-
-    def mla_rope_append_paged_kv_assign_q(
-        self,
-        q: torch.Tensor,
-        latent_cache: torch.Tensor,
-        metadata: DSAtrtllmAttentionMetadata,
-        is_generation: bool = False,
-        **kwargs,
-    ) -> None:
-        if is_generation:
-            cached_token_indptr = metadata.gen_cached_token_indptr
-            kv_indptr = metadata.gen_kv_indptr
-            num_seqs = metadata.num_generations
-            max_seq_len = metadata.max_gen_seq_len
-            block_offsets = metadata.kv_cache_block_offsets[:, metadata.
-                                                            num_contexts:]
-        else:
-            cached_token_indptr = metadata.ctx_cached_token_indptr
-            kv_indptr = metadata.ctx_kv_indptr
-            num_seqs = metadata.num_contexts
-            max_seq_len = metadata.max_ctx_seq_len
-            block_offsets = metadata.kv_cache_block_offsets
-        assert self.is_mla_enable and self.mla_params is not None
-        assert metadata.kv_cache_manager is not None
-
-        sink_token_length = 0
-        beam_width = 1
-
-        torch.ops.trtllm.mla_rope_append_paged_kv_assign_q(
-            q,
-            latent_cache,
-            num_seqs,
-            cached_token_indptr,
-            kv_indptr,
-            max_seq_len,
-            self.wrapper.rotary_cos_sin,
-            self.num_heads,
-            self.mla_params.qk_nope_head_dim,
-            self.mla_params.qk_rope_head_dim,
-            self.mla_params.kv_lora_rank,
-            block_offsets,
-            metadata.kv_cache_manager.kv_cache_pool_pointers,
-            metadata.kv_cache_manager.kv_cache_pool_mapping,
-            self.kv_scale_orig_quant,
-            self.kv_scale_quant_orig,
-            self.get_local_layer_idx(metadata),
-            metadata.kv_cache_manager.tokens_per_block,
-            metadata.kv_cache_manager.max_seq_len,
-            sink_token_length,
-            beam_width,
-            self.wrapper.quant_mode,
-        )
-
-
-class DSACacheManager(KVCacheManager):
-
-    def __init__(
-        self,
-        kv_cache_config: KvCacheConfig,
-        kv_cache_type: CacheTypeCpp,
-        *,
-        num_layers: int,
-        num_kv_heads: Union[int, List[Optional[int]]],
-        head_dim: int,
-        tokens_per_block: int,
-        # Note that max_seq_len is not necessarily equal to kv_cache_config.num_tokens.
-        # It's derived from the model's BuildConfig for consistency with the C++ backend.
-        max_seq_len: int,
-        max_batch_size: int,
-        mapping: Mapping,
-        dtype: DataType = DataType.HALF,
-        spec_config: Optional["DecodingBaseConfig"] = None,
-        layer_mask: Optional[List[bool]] = None,
-        max_num_tokens: int = 8192,
-        model_config: Optional[ModelConfig] = None,
-        max_beam_width: int = 1,
-        sparse_attn_config: "SparseAttentionConfig",
-        **kwargs,
-    ) -> None:
-        self.quant_block_size = 128
-        self.index_head_dim = sparse_attn_config.index_head_dim
-
-        super().__init__(
-            kv_cache_config,
-            kv_cache_type,
-            num_layers=num_layers,
-            num_kv_heads=num_kv_heads,
-            head_dim=head_dim,
-            tokens_per_block=tokens_per_block,
-            max_seq_len=max_seq_len,
-            max_batch_size=max_batch_size,
-            mapping=mapping,
-            dtype=dtype,
-            spec_config=spec_config,
-            layer_mask=layer_mask,
-            max_num_tokens=max_num_tokens,
-            model_config=model_config,
-            max_beam_width=max_beam_width,
-            enable_indexer_k_cache=True,
-            indexer_k_cache_quant_block_size=128,
-            indexer_k_cache_index_head_dim=self.index_head_dim,
-            **kwargs,
-        )
-        self.num_blocks = self.blocks_in_primary_pool
-
-        # Indexer K cache pool for DSA attention
-        # Shape: [num_blocks, self.tokens_per_block * (index_head_dim + scale_size)]
-        # Non-interleaved layout: [fp8_tok0 | fp8_tok1 | ... | scale_tok0 | scale_tok1 | ...]
-        # Store FP8-quantized k values from the indexer
-        self.indexer_k_cache_pool_per_layer = [
-            self.get_indexer_k_cache_pool_data(layer_idx)
-            for layer_idx in range(self.num_local_layers)
-        ]
-
-    def get_indexer_k_cache_buffers(self, layer_idx: int):
-        """Get indexer k cache buffer from a specific layer pool."""
-        block_size = self.tokens_per_block
-        per_token_size = self.index_head_dim + self.index_head_dim // self.quant_block_size * 4
-        layer_offset = self.layer_offsets[layer_idx]
-        return self.indexer_k_cache_pool_per_layer[layer_offset].view(
-            self.num_blocks, block_size, 1, per_token_size)
-
-    def shutdown(self):
-        # Clear Python references BEFORE C++ frees the underlying CUDA buffers
-        self.indexer_k_cache_pool_per_layer = []
-        super().shutdown()
-
-    @staticmethod
-    def get_cache_size_per_token(model_config: ModelConfig,
-                                 mapping: Mapping,
-                                 num_layers: Optional[int] = None,
-                                 **kwargs):
-        config = model_config.pretrained_config
-        sparse_attn_config = model_config.sparse_attention_config
-        index_head_dim = sparse_attn_config.index_head_dim
-        quant_block_size = 128
-
-        # get kv cache dtype bytes
-        mem_per_token = 2
-        quant_config = model_config.quant_config
-        if quant_config is not None and quant_config.quant_mode.has_fp8_kv_cache(
-        ):
-            mem_per_token = 1
-
-        # get head dim
-        head_dim = config.kv_lora_rank + config.qk_rope_head_dim
-
-        num_attention_layers = KVCacheManager._resolve_num_attention_layers(
-            model_config, mapping, num_layers)
-        mem_per_token *= num_attention_layers * head_dim
-
-        # 1 for K, others for indexer K cache
-        head_dim_factor = (index_head_dim +
-                           index_head_dim // quant_block_size * 4) / head_dim
-        kv_factor = 1 + head_dim_factor
-        mem_per_token *= kv_factor
-        return mem_per_token
-
-    def get_cache_bytes_per_token(self):
-        # self.kv_factor for K, others for indexer K cache
-        head_dim_factor = (self.index_head_dim + self.index_head_dim //
-                           self.quant_block_size * 4) / self.head_dim
-        kv_factor = self.kv_factor + head_dim_factor
-        cache_size_per_token = math.ceil(
-            kv_factor * sum(self.num_kv_heads_per_layer) * self.head_dim)
-
-        if self.dtype not in (DataType.FP8, DataType.HALF, DataType.BF16,
-                              DataType.FLOAT, DataType.NVFP4):
-            raise ValueError(f'Cannot support {self.dtype} KV cache.')
-
-        cache_size_bytes_per_token = get_size_in_bytes(cache_size_per_token,
-                                                       self.dtype)
-        if self.dtype == DataType.NVFP4:
-            cache_size_bytes_per_token += self.calculate_scaling_factor_size_bytes(
-                cache_size_per_token,
-                quant_vector_size=16,
-                scaling_factor_dtype=DataType.FP8)
-        return cache_size_bytes_per_token
