+++ tensorrt_llm/_torch/attention_backend/sparse/kernels/__init__.py
@@ -1,0 +1,15 @@
+from .common import (
+    triton_bmm,
+    triton_flatten_to_batch,
+    triton_index_gather,
+    triton_softmax,
+    triton_topk,
+)
+
+__all__ = [
+    "triton_index_gather",
+    "triton_bmm",
+    "triton_softmax",
+    "triton_flatten_to_batch",
+    "triton_topk",
+]
