+++ tensorrt_llm/_torch/attention_backend/sparse/rocket/__init__.py
@@ -1,0 +1,25 @@
+from .backend import RocketTrtllmAttention, RocketVanillaAttention
+from .cache_manager import RocketKVCacheManager
+from .kernels import (
+    triton_rocket_batch_to_flatten,
+    triton_rocket_paged_kt_cache_bmm,
+    triton_rocket_qk_split,
+    triton_rocket_reduce_scores,
+    triton_rocket_update_kt_cache_ctx,
+    triton_rocket_update_kt_cache_gen,
+)
+from .metadata import RocketTrtllmAttentionMetadata, RocketVanillaAttentionMetadata
+
+__all__ = [
+    "RocketTrtllmAttention",
+    "RocketVanillaAttention",
+    "RocketKVCacheManager",
+    "RocketTrtllmAttentionMetadata",
+    "RocketVanillaAttentionMetadata",
+    "triton_rocket_qk_split",
+    "triton_rocket_batch_to_flatten",
+    "triton_rocket_update_kt_cache_gen",
+    "triton_rocket_update_kt_cache_ctx",
+    "triton_rocket_paged_kt_cache_bmm",
+    "triton_rocket_reduce_scores",
+]
