+++ tensorrt_llm/_torch/attention_backend/sparse/utils.py
@@ -1,51 +1,6 @@
-from tensorrt_llm._torch.attention_backend.trtllm import TrtllmAttention
-from tensorrt_llm._torch.pyexecutor.resource_manager import KVCacheManager
-
-from .dsa import DSACacheManager, DSATrtllmAttention
-from .rocket import (RocketKVCacheManager, RocketTrtllmAttention,
-                     RocketVanillaAttention)
-
-
-def get_sparse_attn_kv_cache_manager(
-        sparse_attn_config: "SparseAttentionConfig"):
-    if sparse_attn_config.algorithm == "rocket":
-        return RocketKVCacheManager
-    elif sparse_attn_config.algorithm == "dsa":
-        return DSACacheManager
-    elif sparse_attn_config.algorithm == "skip_softmax":
-        return KVCacheManager
-    else:
-        raise ValueError(
-            f"Unsupported sparse attention algorithm: {sparse_attn_config.algorithm}"
-        )
-
-
-def get_vanilla_sparse_attn_attention_backend(
-        sparse_attn_config: "SparseAttentionConfig"):
-    if sparse_attn_config.algorithm == "rocket":
-        return RocketVanillaAttention
-    else:
-        raise ValueError(
-            f"Unsupported sparse attention algorithm in vanilla attention backend: {sparse_attn_config.algorithm}"
-        )
-
-
-def get_trtllm_sparse_attn_attention_backend(
-        sparse_attn_config: "SparseAttentionConfig"):
-    if sparse_attn_config.algorithm == "rocket":
-        return RocketTrtllmAttention
-    elif sparse_attn_config.algorithm == "dsa":
-        return DSATrtllmAttention
-    elif sparse_attn_config.algorithm == "skip_softmax":
-        return TrtllmAttention
-    else:
-        raise ValueError(
-            f"Unsupported sparse attention algorithm in trtllm attention backend: {sparse_attn_config.algorithm}"
-        )
-
-
-def get_flashinfer_sparse_attn_attention_backend(
-        sparse_attn_config: "SparseAttentionConfig"):
-    raise ValueError(
-        f"Unsupported sparse attention algorithm in flashinfer attention backend: {sparse_attn_config.algorithm}"
-    )
+# Backward-compatibility re-export wrapper.
+# New code should import from .registry instead.
+from .registry import get_flashinfer_sparse_attn_attention_backend  # noqa: F401
+from .registry import get_sparse_attn_kv_cache_manager  # noqa: F401
+from .registry import get_trtllm_sparse_attn_attention_backend  # noqa: F401
+from .registry import get_vanilla_sparse_attn_attention_backend  # noqa: F401
